+++ foo_cover_resizer/CoverRemover.hpp
@@ -8,10 +8,10 @@ public:
 	void run(threaded_process_status& status, abort_callback& abort) override
 	{
 		auto api = file_lock_manager::get();
-		const uint32_t count = m_handles.get_count();
+		const size_t count = m_handles.get_count();
 		std::set<pfc::string8> paths;
 
-		for (const uint32_t i : std::views::iota(0U, count))
+		for (const size_t i : std::views::iota(0U, count))
 		{
 			abort.check();
